util function, minor refactor
