+++ src/window_main/collection.js
@@ -11,7 +11,7 @@ import {
 import db from "../shared/database";
 import pd from "../shared/player-data";
 import { queryElements as $$, createDiv } from "../shared/dom-fns";
-import { createSelect } from "../shared/createSelect";
+import createSelect from "./createSelect";
 import { addCardHover, attachOwnerhipStars } from "../shared/card-hover";
 import {
   collectionSortRarity,
