+++ src/window_main/stats-panel.js
@@ -1,6 +1,6 @@
 import { MANA, RANKS } from "../shared/constants";
 import { createDiv, createLabel } from "../shared/dom-fns";
-import { createSelect } from "../shared/createSelect";
+import createSelect from "./createSelect";
 import { get_rank_index, toDDHHMMSS, toMMSS } from "../shared/util";
 import {
   compareWinrates,
