+++ src/window_main/tournaments.js
@@ -1,6 +1,6 @@
 import anime from "animejs";
 import { queryElements as $$, createDiv } from "../shared/dom-fns";
-import { createSelect } from "../shared/createSelect";
+import createSelect from "./createSelect";
 import * as deckDrawer from "../shared/deck-drawer";
 import { EASING_DEFAULT } from "../shared/constants.js";
 import {
